refactor: rename NavigationItem to VerticalNavigationItem
