+++ src/components/NavigationItem/VerticalNavigationItem.js
